requirements.txt
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
Django==2.2
django-ckeditor==5.9.0
django-debug-toolbar==2.2
django-extensions==3.0.3
django-js-asset==1.2.2
django-mptt==0.11.0
idna==2.10
ipython==7.16.1
ipython-genutils==0.2.0
isort==5.1.4
jedi==0.17.1
parso==0.7.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.5
Pygments==2.6.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.2.5
